The hacker news API has a weird edge case: Some of the returned responses do NOT have a url Property
This one does: https://hacker-news.firebaseio.com/v0/item/39688823.json

This one doesnt
https://hacker-news.firebaseio.com/v0/item/39688937.json

Updated the source script to iterate over responses until one with
a url property is found.

NOTE: Functions has an upper limit to the number of HTTP requests which is
the upper bound of the iterations possible.